requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
feedparser
requests
